services/realtime-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
redis[hiredis]>=5.0.0
kafka-python>=2.0.2
